auth/requirements.txt
Django==1.5.1
py-bcrypt==0.3
wsgiref==0.1.2
